+++ lib/form/addon/sfFormObject.class.php
@@ -10,7 +10,7 @@
 
 /**
  * Base class for forms that deal with a single object.
- * 
+ *
  * @package    symfony
  * @subpackage form
  * @author     Kris Wallsmith <kris.wallsmith@symfony-project.com>
@@ -24,7 +24,7 @@ abstract class sfFormObject extends BaseForm
 
   /**
    * Returns the current model name.
-   * 
+   *
    * @return string
    */
   abstract public function getModelName();
@@ -50,7 +50,7 @@ abstract class sfFormObject extends BaseForm
    * Processes cleaned up values.
    *
    * @param  array $values An array of values
-   * 
+   *
    * @return array An array of cleaned up values
    */
   abstract public function processValues($values);
@@ -108,7 +108,7 @@ abstract class sfFormObject extends BaseForm
    * @return mixed The current saved object
    *
    * @see doSave()
-   * 
+   *
    * @throws sfValidatorError If the form is not valid
    */
   public function save($con = null)
@@ -123,10 +123,9 @@ abstract class sfFormObject extends BaseForm
       $con = $this->getConnection();
     }
 
+    $con->beginTransaction();
     try
     {
-      $con->beginTransaction();
-
       $this->doSave($con);
 
       $con->commit();
@@ -151,17 +150,26 @@ abstract class sfFormObject extends BaseForm
    */
   protected function doSave($con = null)
   {
+    $this->updateObject();
+    $this->saveObject($con);
+  }
+
+  /**
+   * Save form object
+   *
+   * @param  mixed $con An optional connection object
+   */
+  public function saveObject($con = null)
+  {
     if (null === $con)
     {
       $con = $this->getConnection();
     }
 
-    $this->updateObject();
-
     $this->getObject()->save($con);
 
     // embedded forms
-    $this->saveEmbeddedForms($con);
+    $this->saveObjectEmbeddedForms($con);
   }
 
   /**
@@ -225,7 +233,7 @@ abstract class sfFormObject extends BaseForm
    * @param mixed $con   An optional connection object
    * @param array $forms An array of forms
    */
-  public function saveEmbeddedForms($con = null, $forms = null)
+  public function saveObjectEmbeddedForms($con = null, $forms = null)
   {
     if (null === $con)
     {
@@ -241,12 +249,11 @@ abstract class sfFormObject extends BaseForm
     {
       if ($form instanceof sfFormObject)
       {
-        $form->getObject()->save($con);
-        $form->saveEmbeddedForms($con);
+        $form->saveObject($con);
       }
       else
       {
-        $this->saveEmbeddedForms($con, $form->getEmbeddedForms());
+        $this->saveObjectEmbeddedForms($con, $form->getEmbeddedForms());
       }
     }
   }