+++ lib/validator/sfValidatorErrorSchema.class.php
@@ -30,7 +30,7 @@ class sfValidatorErrorSchema extends sfValidatorError implements ArrayAccess, It
    * @param sfValidatorBase $validator  An sfValidatorBase instance
    * @param array           $errors     An array of errors
    */
-  public function __construct(sfValidatorBase $validator, $errors = array())
+  public function __construct(sfValidatorBase $validator)
   {
     $this->validator = $validator;
     $this->arguments = array();
@@ -38,8 +38,6 @@ class sfValidatorErrorSchema extends sfValidatorError implements ArrayAccess, It
     // override default exception message and code
     $this->code    = '';
     $this->message = '';
-
-    $this->addErrors($errors);
   }
 
   /**
@@ -54,7 +52,7 @@ class sfValidatorErrorSchema extends sfValidatorError implements ArrayAccess, It
    */
   public function addError(sfValidatorError $error, $name = null)
   {
-    if (null === $name || is_integer($name))
+    if (null === $name)
     {
       if ($error instanceof sfValidatorErrorSchema)
       {
@@ -66,35 +64,24 @@ class sfValidatorErrorSchema extends sfValidatorError implements ArrayAccess, It
         $this->errors[] = $error;
       }
     }
+    else if (isset($this->namedErrors[$name]))
+    {
+      if (!$this->namedErrors[$name] instanceof sfValidatorErrorSchema)
+      {
+        $current = $this->namedErrors[$name];
+        $this->namedErrors[$name] = new sfValidatorErrorSchema($current->getValidator());
+        $this->namedErrors[$name]->addError($current);
+      }
+
+      $method = $error instanceof sfValidatorErrorSchema ? 'addErrors' : 'addError';
+      $this->namedErrors[$name]->$method($error);
+
+      $this->errors[$name] = $this->namedErrors[$name];
+    }
     else
     {
-      if (!isset($this->namedErrors[$name]) && !$error instanceof sfValidatorErrorSchema)
-      {
-        $this->namedErrors[$name] = $error;
-        $this->errors[$name] = $error;
-      }
-      else
-      {
-        if (!isset($this->namedErrors[$name]))
-        {
-          $this->namedErrors[$name] = new sfValidatorErrorSchema($error->getValidator());
-          $this->errors[$name] = new sfValidatorErrorSchema($error->getValidator());
-        }
-        else if (!$this->namedErrors[$name] instanceof sfValidatorErrorSchema)
-        {
-          $current = $this->namedErrors[$name];
-          $this->namedErrors[$name] = new sfValidatorErrorSchema($current->getValidator());
-          $this->errors[$name] = new sfValidatorErrorSchema($current->getValidator());
-
-          $method = $current instanceof sfValidatorErrorSchema ? 'addErrors' : 'addError';
-          $this->namedErrors[$name]->$method($current);
-          $this->errors[$name]->$method($current);
-        }
-
-        $method = $error instanceof sfValidatorErrorSchema ? 'addErrors' : 'addError';
-        $this->namedErrors[$name]->$method($error);
-        $this->errors[$name]->$method($error);
-      }
+      $this->namedErrors[$name] = $error;
+      $this->errors[$name] = $error;
     }
 
     $this->updateCode();
@@ -104,32 +91,22 @@ class sfValidatorErrorSchema extends sfValidatorError implements ArrayAccess, It
   }
 
   /**
-   * Adds an array of errors.
+   * Adds a collection of errors.
    *
-   * @param array $errors  An array of sfValidatorError instances
+   * @param sfValidatorErrorSchema $errorsAn sfValidatorErrorSchema instance
    *
    * @return sfValidatorErrorSchema The current error schema instance
    */
-  public function addErrors($errors)
+  public function addErrors(sfValidatorErrorSchema $errors)
   {
-    if ($errors instanceof sfValidatorErrorSchema)
+    foreach ($errors->getGlobalErrors() as $error)
     {
-      foreach ($errors->getGlobalErrors() as $error)
-      {
-        $this->addError($error);
-      }
-
-      foreach ($errors->getNamedErrors() as $name => $error)
-      {
-        $this->addError($error, (string) $name);
-      }
+      $this->addError($error);
     }
-    else
+
+    foreach ($errors->getNamedErrors() as $name => $error)
     {
-      foreach ($errors as $name => $error)
-      {
-        $this->addError($error, $name);
-      }
+      $this->addError($error, $name);
     }
 
     return $this;