+++ lib/widget/sfWidgetFormSchema.class.php
@@ -500,9 +500,11 @@ class sfWidgetFormSchema extends sfWidgetForm implements ArrayAccess
       throw new InvalidArgumentException(sprintf('The field named "%s" does not exist.', $name));
     }
 
-    if ($widget instanceof sfWidgetFormSchema && $errors && !$errors instanceof sfValidatorErrorSchema)
+    if ($errors && $widget instanceof sfWidgetFormSchema && !$errors instanceof sfValidatorErrorSchema)
     {
-      $errors = new sfValidatorErrorSchema($errors->getValidator(), array($errors));
+      $schema = new sfValidatorErrorSchema($errors->getValidator());
+      $schema->addError($errors);
+      $errors = $schema;
     }
 
     // we clone the widget because we want to change the id format temporarily
@@ -533,7 +535,7 @@ class sfWidgetFormSchema extends sfWidgetForm implements ArrayAccess
 
     if (!is_array($values) && !$values instanceof ArrayAccess)
     {
-      throw new InvalidArgumentException('You must pass an array of values to render a widget schema');
+      throw new InvalidArgumentException('You must pass an array of values or an instanceof ArrayAccess to render a widget schema');
     }
 
     $formFormat = $this->getFormFormatter();
@@ -559,7 +561,7 @@ class sfWidgetFormSchema extends sfWidgetForm implements ArrayAccess
         $field = $this->renderField($name, $value, $widgetAttributes, $error);
 
         // don't add a label tag and errors if we embed a form schema
-        $label = $widget instanceof sfWidgetFormSchema ? $this->getFormFormatter()->generateLabelName($name) : $this->getFormFormatter()->generateLabel($name);
+        $label = $widget instanceof sfWidgetFormSchema ? $formFormat->generateLabelName($name) : $formFormat->generateLabel($name);
         $error = $widget instanceof sfWidgetFormSchema ? array() : $error;
 
         $rows[] = $formFormat->formatRow($label, $field, $error, $this->getHelp($name));
@@ -580,7 +582,7 @@ class sfWidgetFormSchema extends sfWidgetForm implements ArrayAccess
       $rows[0] = implode("\n", $hiddenRows);
     }
 
-    return $this->getFormFormatter()->formatErrorRow($this->getGlobalErrors($errors)).implode('', $rows);
+    return $formFormat->formatErrorRow($this->getGlobalErrors($errors)).implode('', $rows);
   }
 
   /**